+++ src/Util/templates/datastream.h
@@ -33,7 +33,7 @@ class DataStream
 
 
     /** The current reading position. **/
-    mutable uint32_t nReadPos;
+    mutable uint64_t nReadPos;
 
 
     /** The serialization type. **/
@@ -109,7 +109,7 @@ public:
      *  @param[in] nNewPos The position to set to in the stream.
      *
      **/
-    void SetPos(uint32_t nNewPos) const;
+    void SetPos(uint64_t nNewPos) const;
 
 
     /** SetNull
@@ -154,7 +154,7 @@ public:
      *  @return Returns a reference to the DataStream object.
      *
      **/
-    const DataStream& read(char* pch, uint32_t nSize) const;
+    const DataStream& read(char* pch, uint64_t nSize) const;
 
 
     /** write
@@ -165,7 +165,7 @@ public:
      *  @param[in] nSize The total number of bytes to copy.
      *
      **/
-    DataStream& write(const char* pch, uint32_t nSize);
+    DataStream& write(const char* pch, uint64_t nSize);
 
 
     /** Bytes
@@ -180,7 +180,7 @@ public:
      *  Implement the same reserve functionality to vector.
      *
      **/
-    void reserve(const uint32_t nSize);
+    void reserve(const uint64_t nSize);
 
 
     /** begin
@@ -236,7 +236,7 @@ public:
      *  Get the size of the data stream.
      *
      **/
-    size_t size() const;
+    uint64_t size() const;
 
 
     /** Operator Overload <<
